Finishing moves

+++ Revision2/src/model/Product.java
@@ -1,15 +1,17 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class Product {
     private String name;
-    private Double price;
+    private BigDecimal price;
     private String barcode;
     private Integer remainder;
 
     public Product() {
     }
 
-    public Product(String name, Double price, String barcode, Integer remainder) {
+    public Product(String name, BigDecimal price, String barcode, Integer remainder) {
         this.name = name;
         this.price = price;
         this.barcode = barcode;
@@ -24,11 +26,11 @@ public class Product {
         this.name = name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
